complete notification related pin

+++ src/main/java/com/socialmedia/model/Types.java
@@ -1,23 +1,19 @@
 package com.socialmedia.model;
 
-import java.util.ArrayList;
-import java.util.List;
-
 import jakarta.persistence.Entity;
 import jakarta.persistence.GeneratedValue;
 import jakarta.persistence.GenerationType;
 import jakarta.persistence.Id;
-import jakarta.persistence.OneToMany;
 
 @Entity
 public class Types {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long id;
+    private int id;
     private String typeName;
 
-    public Long getId() {
+    public int getId() {
         return id;
     }
